+++ packages/salesforcedx-vscode-apex/src/constants.ts
@@ -90,6 +90,7 @@ const startPos = new vscode.Position(0, 0);
 const endPos = new vscode.Position(0, 1);
 export const APEX_GROUP_RANGE = new vscode.Range(startPos, endPos);
 export const SET_JAVA_DOC_LINK =
-  'https://forcedotcom.github.io/salesforcedx-vscode/articles/troubleshooting#set-your-java-version';
+  'https://forcedotcom.github.io/salesforcedx-vscode/articles/getting-started/java-setup';
 export const SFDX_APEX_CONFIGURATION_NAME = 'salesforcedx-vscode-apex';
-export const ENABLE_SOBJECT_REFRESH_ON_STARTUP = 'enable-sobject-refresh-on-startup';
+export const ENABLE_SOBJECT_REFRESH_ON_STARTUP =
+  'enable-sobject-refresh-on-startup';