+++ packages/salesforcedx-vscode-core/src/commands/templates/forceApexTriggerCreate.ts
@@ -19,10 +19,12 @@ import {
   SfdxWorkspaceChecker
 } from '../commands';
 import {
-  BaseTemplateCommand,
-  DefaultPathStrategy,
-  FilePathExistsChecker
-} from './baseTemplateCommand';
+  FilePathExistsChecker,
+  GlobStrategyFactory,
+  PathStrategyFactory,
+  SourcePathStrategy
+} from '../util';
+import { BaseTemplateCommand } from './baseTemplateCommand';
 import {
   APEX_TRIGGER_DIRECTORY,
   APEX_TRIGGER_EXTENSION
@@ -39,7 +41,7 @@ export class ForceApexTriggerCreateExecutor extends BaseTemplateCommand {
       .build();
   }
 
-  public sourcePathStrategy = new DefaultPathStrategy();
+  public sourcePathStrategy: SourcePathStrategy = PathStrategyFactory.createDefaultStrategy();
 
   public getDefaultDirectory() {
     return APEX_TRIGGER_DIRECTORY;
@@ -62,9 +64,11 @@ export async function forceApexTriggerCreate() {
     ),
     new ForceApexTriggerCreateExecutor(),
     new FilePathExistsChecker(
-      [APEX_TRIGGER_EXTENSION],
-      new DefaultPathStrategy(),
-      nls.localize('apex_trigger_message_name')
+      GlobStrategyFactory.createCheckFileInGivenPath(APEX_TRIGGER_EXTENSION),
+      nls.localize(
+        'warning_prompt_file_overwrite',
+        nls.localize('apex_trigger_message_name')
+      )
     )
   );
   await commandlet.run();