+++ packages/salesforcedx-vscode-core/src/commands/util/parameterGatherers.ts
@@ -1,0 +1,46 @@
+/*
+ * Copyright (c) 2019, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+import {
+  CancelResponse,
+  ContinueResponse,
+  LocalComponent,
+  ParametersGatherer
+} from '@salesforce/salesforcedx-utils-vscode/out/src/types';
+import { RetrieveDescriber } from '../forceSourceRetrieveMetadata';
+
+export class SimpleGatherer<T> implements ParametersGatherer<T> {
+  private input: T;
+
+  constructor(input: T) {
+    this.input = input;
+  }
+
+  public async gather(): Promise<ContinueResponse<T>> {
+    return {
+      type: 'CONTINUE',
+      data: this.input
+    };
+  }
+}
+
+export class RetrieveComponentOutputGatherer
+  implements ParametersGatherer<LocalComponent[]> {
+  private describer: RetrieveDescriber;
+
+  constructor(describer: RetrieveDescriber) {
+    this.describer = describer;
+  }
+
+  public async gather(): Promise<
+    CancelResponse | ContinueResponse<LocalComponent[]>
+  > {
+    return {
+      type: 'CONTINUE',
+      data: await this.describer.gatherOutputLocations()
+    };
+  }
+}