+++ packages/salesforcedx-vscode-lightning/test/vscode-integration/extension.test.ts
@@ -54,7 +54,11 @@ describe('Test commands', () => {
   let auraExtension: vscode.Extension<any>;
 
   before(async () => {
-    if (vscode.workspace.rootPath) {
+    if (
+      vscode.workspace &&
+      vscode.workspace.workspaceFolders &&
+      vscode.workspace.workspaceFolders[0]
+    ) {
       coreExtension = vscode.extensions.getExtension(
         'salesforce.salesforcedx-vscode-core'
       ) as vscode.Extension<any>;